refactor: core components naming
